backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.13.3
psycopg2-binary==2.9.10
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
ruff==0.6.9
pytest==8.3.2
# Phase 10: Advanced Notification Features
APScheduler==3.10.4
redis==5.0.8
jinja2==3.1.4
python-dateutil==2.9.0
pytz==2024.1
celery==5.4.0
firebase-admin==6.5.0
cryptography==43.0.1
